% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocman-package.R
\docType{package}
\name{biocman-package}
\alias{biocman}
\alias{biocman-package}
\title{biocman: What the Package Does (One Line, Title Case)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Russell Dinnage \email{r.dinnage@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
