% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_decode.R
\name{biocman_decode}
\alias{biocman_decode}
\title{Decode biocman data to bioclim data (19 bioclimatic variables)}
\usage{
biocman_decode(x, standardised = TRUE, device = "cpu")
}
\arguments{
\item{x}{A matrix or data.frame. Columns are assumed to be in order from
BIOMAN1 to BIOMAN5}

\item{standardised}{Should the returned bioclim variables be in standardised
form (the default)? If \code{FALSE} the variables will be back-transformed to the
original scale of the bioclim variables}

\item{device}{A \code{\link[torch:torch_device]{torch::torch_device()}}, on which to run the decoder model, default: \code{"cpu"}}
}
\value{
A matrix of bioclim variables (19 bioclimatic variables). Rows
correspond to row in \code{x}.
}
\description{
Decode biocman data to bioclim data (19 bioclimatic variables)
}
