% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_decode.R
\name{biocman_encode}
\alias{biocman_encode}
\title{Encode bioclim data to biocman data (5 bioclim manifold variables)}
\usage{
biocman_encode(x, standardised = FALSE, device = "cpu")
}
\arguments{
\item{x}{A matrix or data.frame. Columns are assumed to be in order from
BIO01 to BIO19}

\item{standardised}{Are the bioclim variables standardised? If so, this argument
should be \code{TRUE} and they must have been standardised using the mean and SD of the
WorldClim 2.1 2.5 minute data. These values can be found in \code{bioclim_scaling}}

\item{device}{A \code{\link[torch:torch_device]{torch::torch_device()}}, on which to run the encoder model, default: \code{"cpu"}}
}
\value{
A matrix of biocman variables (5 bioclimatic manifold variables). Rows
correspond to row in \code{x}
}
\description{
Encode bioclim data to biocman data (5 bioclim manifold variables)
}
