% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_data}
\alias{get_data}
\title{Return the 5 bioclim manifold variables in a raster format}
\usage{
get_data(return_type = c("rast", "raster", "stars"))
}
\arguments{
\item{return_type}{Return format - One of \code{"rast"} (\code{terra} package: \code{\link[terra:rast]{terra::rast()}}),
\code{"raster"} (\code{raster} package: \code{\link[raster:stack]{raster::stack()}}), or \code{"stars"} (\code{stars} package: \code{\link[stars:st_as_stars]{stars::st_as_stars()}})}
}
\value{
An object of \code{return_type} with the global raster data for biocman
}
\description{
Return the 5 bioclim manifold variables in a raster format
}
\examples{
biocman <- get_data()
## plot 1st manifold variable
plot(biocman[[1]])
}
